% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{buildAnimation}
\alias{buildAnimation}
\title{Build and Cache Animation}
\usage{
buildAnimation(rkey, rtype = "aps")
}
\arguments{
\item{rkey}{unique scan (md5)}

\item{rtype}{file type "aps" or "a3daps" (low / hi resolution)}
}
\value{
nothing
}
\description{
Build and Cache Animation
}
\examples{
buildAnimation("00360f79fd6e02781457eda48f85da90","aps");
buildAnimation("00360f79fd6e02781457eda48f85da90","a3daps");

buildAnimation("0043db5e8c819bffc15261b1f1ac5e42","aps");
buildAnimation("0043db5e8c819bffc15261b1f1ac5e42","a3daps");
}
