% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{buildAnimations}
\alias{buildAnimations}
\title{Build Animations}
\usage{
buildAnimations(rtype = "aps")
}
\arguments{
\item{rtype}{file type "aps" or "a3daps" (low / hi resolution)}
}
\value{
nothing
}
\description{
Build Animations
}
\examples{
buildAnimations("aps");
buildAnimations("a3daps");
}
