% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{buildRawProfileA3DAPS}
\alias{buildRawProfileA3DAPS}
\title{Copy `a3daps` to profiles directory, separate 'data' from 'header' as separate objects}
\usage{
buildRawProfileA3DAPS(a3daps_dir)
}
\arguments{
\item{a3daps_dir}{}
}
\value{
nothing
}
\description{
data.a3daps as raw.Rda
header.a3daps as header.Rda
}
\examples{

buildRawProfileA3DAPS("/data/R_data/_DHS_/a3daps/sample/");
buildRawProfileA3DAPS("/data/R_data/_DHS_/a3daps/stage1/");

}
