% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{loadData}
\alias{loadData}
\title{Load Data from submission, based on type}
\usage{
loadData(rkey, rtype = "a3daps")
}
\arguments{
\item{rkey}{unique scan (md5)}

\item{rtype}{file type "aps" or "a3daps" (low / hi resolution)}
}
\value{
list of two elements $raw and $header
}
\description{
Load Data from submission, based on type
}
\examples{
loadData("0367394485447c1c3485359ba71f52cb","a3daps");
loadData("0367394485447c1c3485359ba71f52cb","aps");
}
