% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPredictions.R
\name{predictMeanProbability}
\alias{predictMeanProbability}
\title{Prediction based on mean of training set}
\usage{
predictMeanProbability()
}
\value{
nothing, stores file in dhsEI/dhsEI/example/trials/submit1
}
\description{
Prediction based on mean of training set
}
\examples{
predictMeanProbability();
}
