% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPredictions.R
\name{predictMeanProbabilityPsych9}
\alias{predictMeanProbabilityPsych9}
\title{Prediction based on mean of training set (with zone-9 offset)}
\usage{
predictMeanProbabilityPsych9(prob = 0.038)
}
\arguments{
\item{prob}{probability for zone-9}
}
\value{
nothing, stores file in dhsEI/dhsEI/example/trials/submit1
}
\description{
Prediction based on mean of training set (with zone-9 offset)
}
\examples{
predictMeanProbabilityPsych9(0.038);
predictMeanProbabilityPsych9(0.05);
}
