% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPredictions.R
\name{predictUniformProbability}
\alias{predictUniformProbability}
\title{Prediction based on some fixed probability}
\usage{
predictUniformProbability(prob = 0.5)
}
\arguments{
\item{prob}{fixed probability [0-1]}
}
\value{
nothing, stores file in dhsEI/dhsEI/example/trials/submit1
}
\description{
Prediction based on some fixed probability
}
\examples{
predictUniformProbability(0.5);
predictUniformProbability(0.15);
predictUniformProbability(0.05);
predictUniformProbability(0.10);
}
