% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_data}
\alias{read_data}
\title{Reads header of `aps` file [also a3daps]}
\usage{
read_data(fname)
}
\arguments{
\item{fname}{string, filename}
}
\value{
array data [unless `ahi` then real/imag]
}
\description{
Reads header of `aps` file [also a3daps]
}
