% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read_header}
\alias{read_header}
\title{Reads header of `aps` file [also a3daps]}
\usage{
read_header(fname)
}
\arguments{
\item{fname}{string, filename}
}
\value{
list of details
}
\description{
Reads header of `aps` file [also a3daps]
}
