% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPredictions.R
\name{writePredictionToCSV}
\alias{writePredictionToCSV}
\title{Write my dataframe format to required output formate}
\usage{
writePredictionToCSV(dframe, dpath = "", sname = "myName")
}
\arguments{
\item{dframe}{my data frame, see data("stage1") for sample}

\item{dpath}{path to write to file}

\item{sname}{stem of file name, not including "CSV"}
}
\value{
string, name of filename stored
}
\description{
Write my dataframe format to required output formate
}
